+++ OvarianCancerSurvival/dataPrepare/resizeImage.py
@@ -15,7 +15,7 @@ nrrdList = glob.glob(nrrdPath + f"/*_CT.nrrd")
 
 for nrrdFile in nrrdList:
     itkImage = sitk.ReadImage(nrrdFile)
-    npVolume = sitk.GetArrayFromImage(itkImage)
+    npVolume = sitk.GetArrayFromImage(itkImage).astype(float)
     S,H,W = npVolume.shape
 
     resultVolume = np.zeros((S,dH,dW),dtype=float)
